requirements.txt
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.7
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.4.2
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
Flask==2.3.3
flask-cors==5.0.1
idna==3.10
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pip==24.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pymongo==4.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.3.7
yarg==0.1.9
